chore: Update Guide.md to link a working static server package (#3049)
The static server package does not seem to work in the later version of React Native. Or, there's a way to make it work but not discussed.
A fork of this [library](https://github.com/birdofpreyru/react-native-static-server) works as expected.
While it may not have many 'stars' the fact that it works and is also maintained, makes it a more suitable candidate.

(cherry picked from commit 2060bd64a36ad8c2242b4d9cabe4e12f3e79f487)